requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Storage (for future S3 integration)
boto3==1.35.95

# Utilities
python-slugify==8.0.4
pillow==10.4.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
